Move the tokenizer related tests to TokenizerTest
