+++ smithy-model/src/test/java/software/amazon/smithy/model/loader/TaggedStringLiteralTest.java
@@ -292,121 +292,7 @@ public class TaggedStringLiteralTest {
         assertThat(TaggedStringLiteral.hasHandler("unknown"), is(false));
     }
 
-    // --- Tokenizer-level tests ---
-
-    private static IdlInternalTokenizer tokenizer(String model) {
-        return new IdlInternalTokenizer("a.smithy", model);
-    }
-
-    @Test
-    public void tokenizesReTagAsTagPlusRawString() {
-        IdlInternalTokenizer t = tokenizer("#re \"^\\d{5}$\"");
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        assertThat(t.getCurrentTokenLexeme().toString(), equalTo("#re"));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_STRING));
-        assertThat(t.getCurrentTokenStringSlice().toString(), equalTo("^\\d{5}$"));
-    }
-
-    @Test
-    public void tokenizesReTaggedTextBlock() {
-        String model = "#re \"\"\"\n"
-                + "    ^\\d{5}$\n"
-                + "    \"\"\"";
-        IdlInternalTokenizer t = tokenizer(model);
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        assertThat(t.getCurrentTokenLexeme().toString(), equalTo("#re"));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_TEXT_BLOCK));
-    }
-
-    @Test
-    public void tokenizesBTagAsTagPlusRawString() {
-        IdlInternalTokenizer t = tokenizer("#b \"Hello world\"");
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        assertThat(t.getCurrentTokenLexeme().toString(), equalTo("#b"));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_STRING));
-        assertThat(t.getCurrentTokenStringSlice().toString(), equalTo("Hello world"));
-    }
-
-    @Test
-    public void tokenizesHexTextBlock() {
-        String model = "#hex \"\"\"\n"
-                + "    81 # array(1)\n"
-                + "    \"\"\"";
-        IdlInternalTokenizer t = tokenizer(model);
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        assertThat(t.getCurrentTokenLexeme().toString(), equalTo("#hex"));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_TEXT_BLOCK));
-    }
-
-    @Test
-    public void tokenizesTimestampTag() {
-        IdlInternalTokenizer t = tokenizer("#timestamp \"2024-01-01T00:00:00Z\"");
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        assertThat(t.getCurrentTokenLexeme().toString(), equalTo("#timestamp"));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_STRING));
-        assertThat(t.getCurrentTokenStringSlice().toString(), equalTo("2024-01-01T00:00:00Z"));
-    }
-
-    // --- Disambiguation ---
-
-    @Test
-    public void poundStillWorksAsToken() {
-        IdlInternalTokenizer t = tokenizer("#");
-        assertThat(t.getCurrentToken(), is(IdlToken.POUND));
-    }
-
-    @Test
-    public void poundFollowedByIdentifierWithoutString() {
-        IdlInternalTokenizer t = tokenizer("#re foo");
-        assertThat(t.getCurrentToken(), is(IdlToken.POUND));
-    }
-
-    @Test
-    public void poundFollowedByNonIdentifier() {
-        IdlInternalTokenizer t = tokenizer("#123");
-        assertThat(t.getCurrentToken(), is(IdlToken.POUND));
-    }
-
-    // --- Raw string escape handling ---
-
-    @Test
-    public void rawStringPreservesEscapedQuote() {
-        // #re "a\"b" — the \" keeps the quote from terminating the string,
-        // raw content includes the backslash and quote.
-        IdlInternalTokenizer t = tokenizer("#re \"a\\\"b\"");
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_STRING));
-        assertThat(t.getCurrentTokenStringSlice().toString(), equalTo("a\\\"b"));
-    }
-
-    @Test
-    public void rawStringPreservesBackslashAtEnd() {
-        // #re "foo\\" — escaped backslash before closing quote
-        IdlInternalTokenizer t = tokenizer("#re \"foo\\\\\"");
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_STRING));
-        assertThat(t.getCurrentTokenStringSlice().toString(), equalTo("foo\\\\"));
-    }
-
-    @Test
-    public void rawTextBlockAllowsUnescapedQuotes() {
-        // In text blocks, lone " doesn't need escaping.
-        String model = "#re \"\"\"\n    a\"b\n    \"\"\"";
-        IdlInternalTokenizer t = tokenizer(model);
-        assertThat(t.getCurrentToken(), is(IdlToken.TAG));
-        t.next();
-        assertThat(t.getCurrentToken(), is(IdlToken.RAW_TEXT_BLOCK));
-        assertThat(t.getCurrentTokenStringSlice().toString(), containsString("a\"b"));
-    }
-
-    // --- Helpers ---
+`    // --- Helpers ---
 
     private static String base64(String s) {
         return Base64.getEncoder().encodeToString(s.getBytes(StandardCharsets.UTF_8));
